Separated the variant classes from the description module.
